php: class users converted to classes Users(user entity) and ExampleUsers(users-collection entity)
